+++ src/Admin/AbstractAdmin.php
@@ -26,6 +26,7 @@ use Sonata\AdminBundle\Datagrid\DatagridMapper;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Datagrid\Pager;
 use Sonata\AdminBundle\Datagrid\ProxyQueryInterface;
+use Sonata\AdminBundle\Exporter\DataSourceInterface;
 use Sonata\AdminBundle\Filter\Persister\FilterPersisterInterface;
 use Sonata\AdminBundle\Form\FormMapper;
 use Sonata\AdminBundle\Form\Type\ModelHiddenType;
@@ -45,7 +46,6 @@ use Sonata\Exporter\Source\SourceIteratorInterface;
 use Sonata\Form\Validator\Constraints\InlineConstraint;
 use Sonata\Form\Validator\ErrorElement;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
-use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -241,6 +241,13 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface, A
      * @var ModelManagerInterface
      */
     protected $modelManager;
+
+    /**
+     * NEXT_MAJOR: Restrict to DataSourceInterface.
+     *
+     * @var DataSourceInterface|null
+     */
+    protected $dataSource;
 
     /**
      * The current request object.
@@ -568,6 +575,19 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface, A
                 $fields[$transLabel] = $field;
             }
         }
+
+        if ($this->getDataSource()) {
+            $query = $datagrid->getQuery();
+
+            return $this->getDataSource()->createIterator($query, $fields);
+        }
+
+        @trigger_error(sprintf(
+            'Using "%s()" without setting a "%s" instance in the admin is deprecated since sonata-admin/admin-bundle 3.79'
+            .' and won\'t be possible in 4.0.',
+            __METHOD__,
+            DataSourceInterface::class
+        ), E_USER_DEPRECATED);
 
         return $this->getModelManager()->getDataSourceIterator($datagrid, $fields);
     }
@@ -1931,6 +1951,19 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface, A
     public function setModelManager(?ModelManagerInterface $modelManager): void
     {
         $this->modelManager = $modelManager;
+    }
+
+    /**
+     * NEXT_MAJOR: Change typehint for DataSourceInterface.
+     */
+    public function getDataSource(): ?DataSourceInterface
+    {
+        return $this->dataSource;
+    }
+
+    public function setDataSource(DataSourceInterface $dataSource)
+    {
+        $this->dataSource = $dataSource;
     }
 
     public function getManagerType(): ?string