+++ src/Templating/TemplateRegistry.php
@@ -15,76 +15,38 @@ namespace Sonata\AdminBundle\Templating;
 
 /**
  * @author Timo Bakx <timobakx@gmail.com>
+ *
+ * NEXT_MAJOR: remove `MutableTemplateRegistryInterface` implementation.
  */
-final class TemplateRegistry implements MutableTemplateRegistryInterface
+final class TemplateRegistry extends AbstractTemplateRegistry implements MutableTemplateRegistryInterface
 {
-    public const TYPE_ARRAY = 'array';
-    public const TYPE_BOOLEAN = 'boolean';
-    public const TYPE_DATE = 'date';
-    public const TYPE_TIME = 'time';
-    public const TYPE_DATETIME = 'datetime';
-    public const TYPE_TEXTAREA = 'textarea';
-    public const TYPE_EMAIL = 'email';
-    public const TYPE_TRANS = 'trans';
-    public const TYPE_STRING = 'string';
-    public const TYPE_INTEGER = 'integer';
-    public const TYPE_FLOAT = 'float';
-    public const TYPE_IDENTIFIER = 'identifier';
-    public const TYPE_CURRENCY = 'currency';
-    public const TYPE_PERCENT = 'percent';
-    public const TYPE_CHOICE = 'choice';
-    public const TYPE_URL = 'url';
-    public const TYPE_HTML = 'html';
-    public const TYPE_MANY_TO_MANY = 'many_to_many';
-    public const TYPE_MANY_TO_ONE = 'many_to_one';
-    public const TYPE_ONE_TO_MANY = 'one_to_many';
-    public const TYPE_ONE_TO_ONE = 'one_to_one';
-
     /**
-     * @var array<string, string>
-     */
-    private $templates = [];
-
-    /**
-     * @param string[] $templates
-     */
-    public function __construct(array $templates = [])
-    {
-        $this->templates = $templates;
-    }
-
-    public function getTemplates(): array
-    {
-        return $this->templates;
-    }
-
-    /**
-     * @param string[] $templates
+     * NEXT_MAJOR: remove this method.
+     *
+     * @deprecated since version sonata-project/admin-bundle 3.39.0 and will be removed in 4.0. Use Sonata\AdminBundle\Templating\MutableTemplateRegistry instead.
      */
     public function setTemplates(array $templates): void
     {
         $this->templates = $templates;
+
+        @trigger_error(sprintf(
+            'Method "%s()" is deprecated since sonata-admin/admin-bundle 3.39 and will be removed in 4.0.',
+            __METHOD__
+        ), E_USER_DEPRECATED);
     }
 
-    public function hasTemplate(string $name): bool
-    {
-        return isset($this->templates[$name]);
-    }
-
-    public function getTemplate(string $name): string
-    {
-        if (isset($this->templates[$name])) {
-            return $this->templates[$name];
-        }
-
-        throw new \InvalidArgumentException(sprintf(
-            'Template named "%s" doesn\'t exist.',
-            $name
-        ));
-    }
-
+    /**
+     * NEXT_MAJOR: remove this method.
+     *
+     * @deprecated since version sonata-project/admin-bundle 3.39.0 and will be removed in 4.0. Use Sonata\AdminBundle\Templating\MutableTemplateRegistry instead.
+     */
     public function setTemplate(string $name, string $template): void
     {
         $this->templates[$name] = $template;
+
+        @trigger_error(sprintf(
+            'Method "%s()" is deprecated since sonata-admin/admin-bundle 3.39 and will be removed in 4.0.',
+            __METHOD__
+        ), E_USER_DEPRECATED);
     }
 }