+++ tests/App/Model/ModelManager.php
@@ -117,6 +117,9 @@ class ModelManager implements ModelManagerInterface, LockInterface
         throw new \BadMethodCallException('Not implemented.');
     }
 
+    /**
+     * NEXT_MAJOR: Remove this method.
+     */
     public function modelTransform(string $class, object $instance): object
     {
         throw new \BadMethodCallException('Not implemented.');