+++ tests/Datagrid/DatagridMapperTest.php
@@ -26,6 +26,7 @@ use Sonata\AdminBundle\Datagrid\ProxyQueryInterface;
 use Sonata\AdminBundle\Filter\Filter;
 use Sonata\AdminBundle\Filter\FilterInterface;
 use Sonata\AdminBundle\Model\ModelManagerInterface;
+use Sonata\AdminBundle\Translator\LabelTranslatorStrategyInterface;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilder;
 
@@ -102,6 +103,17 @@ class DatagridMapperTest extends TestCase
                 return self::DEFAULT_GRANTED_ROLE === $name;
             });
 
+        $labelTranslatorStrategy = $this->createStub(LabelTranslatorStrategyInterface::class);
+        $labelTranslatorStrategy->method('getLabel')->willReturnCallback(
+            static function ($label, $context = '', $type = ''): string {
+                return sprintf('%s.%s_%s', $context, $type, $label);
+            }
+        );
+
+        $admin
+            ->method('getLabelTranslatorStrategy')
+            ->willReturn($labelTranslatorStrategy);
+
         $this->datagridMapper = new DatagridMapper($datagridBuilder, $this->datagrid, $admin);
     }
 
@@ -175,10 +187,11 @@ class DatagridMapperTest extends TestCase
 
         $this->assertTrue($this->datagridMapper->has('fooName'));
 
-        $fieldDescription = $this->datagridMapper->get('fooName');
+        $filter = $this->datagridMapper->get('fooName');
 
-        $this->assertInstanceOf(FilterInterface::class, $fieldDescription);
-        $this->assertSame('fooName', $fieldDescription->getName());
+        $this->assertInstanceOf(FilterInterface::class, $filter);
+        $this->assertSame('fooName', $filter->getName());
+        $this->assertSame('filter.label_fooName', $filter->getLabel());
     }
 
     public function testAddWithoutFieldName(): void
@@ -187,11 +200,11 @@ class DatagridMapperTest extends TestCase
 
         $this->assertTrue($this->datagridMapper->has('foo.bar'));
 
-        $fieldDescription = $this->datagridMapper->get('foo.bar');
+        $filter = $this->datagridMapper->get('foo.bar');
 
-        $this->assertInstanceOf(FilterInterface::class, $fieldDescription);
-        $this->assertSame('foo.bar', $fieldDescription->getName());
-        $this->assertSame('bar', $fieldDescription->getOption('field_name'));
+        $this->assertInstanceOf(FilterInterface::class, $filter);
+        $this->assertSame('foo.bar', $filter->getName());
+        $this->assertSame('bar', $filter->getOption('field_name'));
     }
 
     public function testAddRemove(): void